need more structure to grouped datasets
